Inline `JaxTDigest` into `tdigest_normalizer` and remove its standalone implementation.
